Update pyfunc test to include retry
